split main.dart into multiple files
