+++ lib/main.dart
@@ -1,7 +1,8 @@
-import 'package:english_words/english_words.dart';
 import 'package:flutter/material.dart';
+import 'package:english_words/english_words.dart';
 import 'package:provider/provider.dart';
-import 'package:table_calendar/table_calendar.dart';
+import 'pages/sign_up_page.dart';
+import 'pages/home_page.dart';
 
 void main() => runApp(const SignUpApp());
 
@@ -45,614 +46,5 @@ class MyAppState extends ChangeNotifier {
 
   void addNote() {
     // TODO
-  }
-}
-
-//Start of Sign Up Page
-class SignUpApp extends StatelessWidget {
-  const SignUpApp();
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      routes: {
-        '/': (context) => const SignUpScreen(),
-        '/welcome': (context) => const MyApp(),
-      },
-    );
-  }
-}
-
-//Sign Up Background
-class SignUpScreen extends StatelessWidget {
-  const SignUpScreen();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.grey[200],
-      body: const Center(
-        child: SizedBox(
-          width: 400,
-          child: Card(
-            child: SignUpForm(),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-//The Sign Up Box State
-class SignUpForm extends StatefulWidget {
-  const SignUpForm();
-
-  @override
-  State<SignUpForm> createState() => _SignUpFormState();
-}
-
-//Text Box
-class _SignUpFormState extends State<SignUpForm> {
-  final _firstNameTextController = TextEditingController();
-  final _lastNameTextController = TextEditingController();
-  final _usernameTextController = TextEditingController();
-
-  double _formProgress = 0;
-
-  void _showWelcomeScreen() {
-    Navigator.of(context).pushNamed('/welcome');
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          LinearProgressIndicator(value: _formProgress),
-          Text('Sign up', style: Theme.of(context).textTheme.headlineMedium),
-          Padding(
-            padding: const EdgeInsets.all(8),
-            child: TextFormField(
-              controller: _firstNameTextController,
-              decoration: const InputDecoration(hintText: 'First name'),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8),
-            child: TextFormField(
-              controller: _lastNameTextController,
-              decoration: const InputDecoration(hintText: 'Last name'),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8),
-            child: TextFormField(
-              controller: _usernameTextController,
-              decoration: const InputDecoration(hintText: 'Username'),
-            ),
-          ),
-          TextButton(
-            style: ButtonStyle(
-              foregroundColor: MaterialStateProperty.resolveWith(
-                  (Set<MaterialState> states) {
-                return states.contains(MaterialState.disabled)
-                    ? null
-                    : Colors.white;
-              }),
-              backgroundColor: MaterialStateProperty.resolveWith(
-                  (Set<MaterialState> states) {
-                return states.contains(MaterialState.disabled)
-                    ? null
-                    : Colors.blue;
-              }),
-            ),
-            onPressed: _showWelcomeScreen,
-            child: const Text('Sign up'),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// Sends User to App after sign up
-class MyHomePage extends StatefulWidget {
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  var selectedIndex = 0;
-
-  @override
-  Widget build(BuildContext context) {
-    Widget page;
-    //Do we need brakes for each case?
-    switch (selectedIndex) {
-      case 0:
-        page = GeneratorPage();
-      case 1:
-        page = FavoritesPage();
-      case 2:
-        page = NewNotePage();
-      case 3:
-        page = SavedNotesPage();
-      case 4:
-        page = CalendarPage();
-      default:
-        throw UnimplementedError('no widget for $selectedIndex');
-    }
-
-    return LayoutBuilder(builder: (context, constraints) {
-      return Scaffold(
-        body: Row(
-          children: [
-            SafeArea(
-              child: NavigationRail(
-                extended: constraints.maxWidth >= 600,
-                destinations: [
-                  NavigationRailDestination(
-                    icon: Icon(Icons.home),
-                    label: Text('Home'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.favorite),
-                    label: Text('Favorite Notes'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.add),
-                    label: Text('New Note'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.notes),
-                    label: Text('View Notes'),
-                  ),
-                  NavigationRailDestination(
-                    icon: Icon(Icons.calendar_month),
-                    label: Text('View Calendar'),
-                  ),
-                ],
-                selectedIndex: selectedIndex,
-                onDestinationSelected: (value) {
-                  setState(() {
-                    selectedIndex = value;
-                  });
-                },
-              ),
-            ),
-            Expanded(
-              child: Container(
-                color: Theme.of(context).colorScheme.primaryContainer,
-                child: page,
-              ),
-            ),
-          ],
-        ),
-      );
-    });
-  }
-}
-
-class GeneratorPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    var appState = context.watch<MyAppState>();
-    var pair = appState.current;
-
-    IconData icon;
-    if (appState.favorites.contains(pair)) {
-      icon = Icons.favorite;
-    } else {
-      icon = Icons.favorite_border;
-    }
-
-    return MaterialApp(
-      home: Scaffold(
-        body: Container(
-          padding: EdgeInsets.only(top: 100),
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              //Upcoming Events Tab
-              Column(
-                children: [
-                  Row(
-                    children: [
-                      Container(
-                        width: 50,
-                        height: 50,
-                        decoration: BoxDecoration(color: Colors.amberAccent),
-                        child: Icon(Icons.calendar_month),
-                      ),
-                      Text('Upcoming Events')
-                    ],
-                  )
-                ],
-              ),
-              //Previous Note Tab
-              Column(
-                children: [
-                  Row(
-                    children: [
-                      Container(
-                        width: 50,
-                        height: 50,
-                        decoration: BoxDecoration(color: Colors.amberAccent),
-                        child: Icon(Icons.note),
-                      ),
-                      Text('Previous Note')
-                    ],
-                  )
-                ],
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class BigCard extends StatelessWidget {
-  const BigCard({
-    super.key,
-    required this.pair,
-  });
-
-  final WordPair pair;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final style = theme.textTheme.displayMedium!.copyWith(
-      color: theme.colorScheme.onPrimary,
-    );
-
-    return Card(
-      color: theme.colorScheme.primary,
-      child: Padding(
-        padding: const EdgeInsets.all(20),
-        child: Text(
-          pair.asLowerCase,
-          style: style,
-          semanticsLabel: "${pair.first} ${pair.second}",
-        ),
-      ),
-    );
-  }
-}
-
-class FavoritesPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    var appState = context.watch<MyAppState>();
-
-    if (appState.favorites.isEmpty) {
-      return Center(
-        child: Text('No favorites yet.'),
-      );
-    }
-
-    return ListView(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(20),
-          child: Text('You have '
-              '${appState.favorites.length} favorites:'),
-        ),
-        for (var pair in appState.favorites)
-          ListTile(
-            leading: Icon(Icons.favorite),
-            title: Text(pair.asLowerCase),
-          ),
-      ],
-    );
-  }
-}
-
-//Do we need A page to create notes, could just be a button on a note page
-class NewNotePage extends StatefulWidget {
-  const NewNotePage({super.key});
-  @override
-  State<NewNotePage> createState() => _NewNotePageState();
-}
-
-class _NewNotePageState extends State<NewNotePage> {
-  late TextEditingController _controller;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = TextEditingController();
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          children: [
-            TextField(
-              controller: _controller,
-              onSubmitted: (String value) async {
-                await showDialog<void>(
-                  context: context,
-                  builder: (BuildContext context) {
-                    return AlertDialog(
-                      title: const Text('Thanks!'),
-                      content: Text(
-                          'You typed "$value", which has length ${value.characters.length}.'),
-                      actions: <Widget>[
-                        TextButton(
-                          onPressed: () {
-                            Navigator.pop(context);
-                          },
-                          child: const Text('OK'),
-                        ),
-                      ],
-                    );
-                  },
-                );
-              },
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class SavedNotesPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        body: Container(
-          padding: EdgeInsets.only(top: 100),
-          child: Column(
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  const NoteCard(),
-                  const NoteCard(),
-                  const NoteCard()
-                ],
-              )
-            ],
-          ),
-        ),
-        floatingActionButton: FloatingActionButton(
-          tooltip: 'New Note',
-          onPressed: () {},
-          child: Icon(Icons.add),
-        ),
-        appBar: AppBar(
-          title: Text('Notes'),
-          backgroundColor: const Color.fromARGB(255, 255, 195, 177),
-        ),
-      ),
-    );
-  }
-}
-
-//Creates a clickable note card
-class NoteCard extends StatelessWidget {
-  const NoteCard({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Card(
-        clipBehavior: Clip.hardEdge,
-        child: InkWell(
-          splashColor: Colors.blueAccent,
-          //Clicking Function
-          onTap: () {
-            debugPrint('Note Tapped');
-          },
-          child: const SizedBox(
-            width: 200,
-            height: 100,
-            child: Text(
-              'Your Note Here',
-              textAlign: TextAlign.center,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class CalendarPage extends StatefulWidget {
-  const CalendarPage({super.key});
-  @override
-  State<CalendarPage> createState() => _CalendarPageState();
-}
-
-class _CalendarPageState extends State<CalendarPage> {
-  //Limits
-  final kFirstDay = DateTime.utc(2021, 1, 1);
-  final kLastDay = DateTime.utc(2030, 12, 31);
-
-  //Formating
-  CalendarFormat _calendarFormat = CalendarFormat.month;
-  DateTime _focusedDay = DateTime.now();
-  DateTime? _selectedDay;
-
-  //Add Event Function
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Your Calendar'),
-      ),
-      //"+" button on calendar page, will eventually be used to add events
-      floatingActionButton: FloatingActionButton(
-        onPressed: () => _addEventDialog(context),
-        tooltip: 'Add Event',
-        child: Icon(Icons.add),
-      ),
-      body: TableCalendar(
-        firstDay: kFirstDay,
-        lastDay: kLastDay,
-        focusedDay: _focusedDay,
-        calendarFormat: _calendarFormat,
-        selectedDayPredicate: (day) {
-          // Use `selectedDayPredicate` to determine which day is currently selected.
-          // If this returns true, then `day` will be marked as selected.
-
-          // Using `isSameDay` is recommended to disregard
-          // the time-part of compared DateTime objects.
-          return isSameDay(_selectedDay, day);
-        },
-        onDaySelected: (selectedDay, focusedDay) {
-          if (!isSameDay(_selectedDay, selectedDay)) {
-            // Call `setState()` when updating the selected day
-            setState(() {
-              _selectedDay = selectedDay;
-              _focusedDay = focusedDay;
-            });
-          }
-        },
-        onFormatChanged: (format) {
-          if (_calendarFormat != format) {
-            // Call `setState()` when updating calendar format
-            setState(() {
-              _calendarFormat = format;
-            });
-          }
-        },
-        onPageChanged: (focusedDay) {
-          // No need to call `setState()` here
-          _focusedDay = focusedDay;
-        },
-      ),
-    );
-  }
-
-  Future<void> _addEventDialog(BuildContext context) {
-    DateTime selectedDate = DateTime.now();
-    TimeOfDay selectedTime = TimeOfDay.now();
-    TextEditingController startDateController = TextEditingController();
-    // TextEditingController endDateController = TextEditingController();
-    TextEditingController startTimeController = TextEditingController();
-    // TextEditingController endTimeController = TextEditingController();
-    startDateController.text = "${selectedDate.toLocal()}".split(' ')[0];
-    // endDateController.text = "${selectedDate.toLocal()}".split(' ')[0];
-    startTimeController.text = "${selectedTime.hour}:${selectedTime.minute}";
-
-    Future<void> selectDate(BuildContext context) async {
-      final DateTime? picked = await showDatePicker(
-          context: context,
-          initialDate: selectedDate,
-          firstDate: DateTime(2021),
-          lastDate: DateTime(2030));
-      if (picked != null && picked != selectedDate) {
-        setState(() {
-          selectedDate = picked;
-          startDateController.text = "${selectedDate.toLocal()}".split(' ')[0];
-        });
-      }
-    }
-
-    Future<void> selectTime(BuildContext context) async {
-      final TimeOfDay? picked = await showTimePicker(
-        context: context,
-        initialTime: TimeOfDay.now(),
-      );
-      if (picked != null && picked != selectedTime) {
-        setState(() {
-          selectedTime = picked;
-          startTimeController.text =
-              "${selectedTime.hour}:${selectedTime.minute}";
-        });
-      }
-    }
-
-    return showAdaptiveDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return Dialog(
-          shape: (RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(20.0),
-          )),
-          // title: const Text('Add Event'),
-          child: Container(
-            height: 300,
-            width: 300,
-            padding: EdgeInsets.all(16),
-            child: Column(
-              children: [
-                TextField(
-                  decoration: const InputDecoration(
-                    hintText: 'Event Name',
-                  ),
-                ),
-                TextField(
-                  // expands: true,
-                  // maxLines: null,
-                  // minLines: null,
-                  decoration: const InputDecoration(
-                    hintText: 'Event Description',
-                  ),
-                ),
-                Row(
-                  children: [
-                    Expanded(
-                        child: Center(
-                            child: TextField(
-                      controller: startDateController,
-                      readOnly: true,
-                      decoration: const InputDecoration(
-                        hintText: 'Event Date',
-                      ),
-                    ))),
-                    IconButton(
-                      onPressed: () => selectDate(context),
-                      icon: Icon(Icons.calendar_today),
-                    ),
-                  ],
-                ),
-                Row(
-                  children: [
-                    Expanded(
-                        child: Center(
-                      child: TextField(
-                        controller: startTimeController,
-                        readOnly: true,
-                        decoration: const InputDecoration(
-                          hintText: 'Event Time',
-                        ),
-                      ),
-                    )),
-                    IconButton(
-                      onPressed: () => selectTime(context),
-                      icon: Icon(Icons.access_time),
-                    ),
-                  ],
-                ),
-                Spacer(),
-                ElevatedButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                  child: const Text('Add'),
-                ),
-              ],
-            ),
-          ),
-        );
-      },
-    );
   }
 }
