+++ lib/pages/generator_page.dart
@@ -1,0 +1,93 @@
+import 'package:flutter/material.dart';
+import 'package:english_words/english_words.dart';
+import '../main.dart';
+import 'package:provider/provider.dart';
+
+class GeneratorPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    var appState = context.watch<MyAppState>();
+    var pair = appState.current;
+
+    IconData icon;
+    if (appState.favorites.contains(pair)) {
+      icon = Icons.favorite;
+    } else {
+      icon = Icons.favorite_border;
+    }
+
+    return MaterialApp(
+      home: Scaffold(
+        body: Container(
+          padding: EdgeInsets.only(top: 100),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceAround,
+            children: [
+              //Upcoming Events Tab
+              Column(
+                children: [
+                  Row(
+                    children: [
+                      Container(
+                        width: 50,
+                        height: 50,
+                        decoration: BoxDecoration(color: Colors.amberAccent),
+                        child: Icon(Icons.calendar_month),
+                      ),
+                      Text('Upcoming Events')
+                    ],
+                  )
+                ],
+              ),
+              //Previous Note Tab
+              Column(
+                children: [
+                  Row(
+                    children: [
+                      Container(
+                        width: 50,
+                        height: 50,
+                        decoration: BoxDecoration(color: Colors.amberAccent),
+                        child: Icon(Icons.note),
+                      ),
+                      Text('Previous Note')
+                    ],
+                  )
+                ],
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class BigCard extends StatelessWidget {
+  const BigCard({
+    super.key,
+    required this.pair,
+  });
+
+  final WordPair pair;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    final style = theme.textTheme.displayMedium!.copyWith(
+      color: theme.colorScheme.onPrimary,
+    );
+
+    return Card(
+      color: theme.colorScheme.primary,
+      child: Padding(
+        padding: const EdgeInsets.all(20),
+        child: Text(
+          pair.asLowerCase,
+          style: style,
+          semanticsLabel: "${pair.first} ${pair.second}",
+        ),
+      ),
+    );
+  }
+}
