+++ lib/pages/home_page.dart
@@ -1,0 +1,83 @@
+import 'package:flutter/material.dart';
+import '/pages/generator_page.dart';
+import '/pages/fav_page.dart';
+import '/pages/new_note_page.dart';
+import '/pages/saved_notes_page.dart';
+import '/pages/calendar_page.dart';
+
+class MyHomePage extends StatefulWidget {
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  var selectedIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    Widget page;
+    //Do we need brakes for each case?
+    switch (selectedIndex) {
+      case 0:
+        page = GeneratorPage();
+      case 1:
+        page = FavoritesPage();
+      case 2:
+        page = NewNotePage();
+      case 3:
+        page = SavedNotesPage();
+      case 4:
+        page = CalendarPage();
+      default:
+        throw UnimplementedError('no widget for $selectedIndex');
+    }
+
+    return LayoutBuilder(builder: (context, constraints) {
+      return Scaffold(
+        body: Row(
+          children: [
+            SafeArea(
+              child: NavigationRail(
+                extended: constraints.maxWidth >= 600,
+                destinations: [
+                  NavigationRailDestination(
+                    icon: Icon(Icons.home),
+                    label: Text('Home'),
+                  ),
+                  NavigationRailDestination(
+                    icon: Icon(Icons.favorite),
+                    label: Text('Favorite Notes'),
+                  ),
+                  NavigationRailDestination(
+                    icon: Icon(Icons.add),
+                    label: Text('New Note'),
+                  ),
+                  NavigationRailDestination(
+                    icon: Icon(Icons.notes),
+                    label: Text('View Notes'),
+                  ),
+                  NavigationRailDestination(
+                    icon: Icon(Icons.calendar_month),
+                    label: Text('View Calendar'),
+                  ),
+                ],
+                selectedIndex: selectedIndex,
+                onDestinationSelected: (value) {
+                  setState(() {
+                    selectedIndex = value;
+                  });
+                },
+              ),
+            ),
+            Expanded(
+              child: Container(
+                color: Theme.of(context).colorScheme.primaryContainer,
+                child: page,
+              ),
+            ),
+          ],
+        ),
+      );
+    });
+  }
+}
