+++ lib/pages/new_note_page.dart
@@ -1,0 +1,58 @@
+import 'package:flutter/material.dart';
+
+class NewNotePage extends StatefulWidget {
+  const NewNotePage({super.key});
+  @override
+  State<NewNotePage> createState() => _NewNotePageState();
+}
+
+class _NewNotePageState extends State<NewNotePage> {
+  late TextEditingController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = TextEditingController();
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Column(
+          children: [
+            TextField(
+              controller: _controller,
+              onSubmitted: (String value) async {
+                await showDialog<void>(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: const Text('Thanks!'),
+                      content: Text(
+                          'You typed "$value", which has length ${value.characters.length}.'),
+                      actions: <Widget>[
+                        TextButton(
+                          onPressed: () {
+                            Navigator.pop(context);
+                          },
+                          child: const Text('OK'),
+                        ),
+                      ],
+                    );
+                  },
+                );
+              },
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
