+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,7 +5,7 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = codelab
+PRODUCT_NAME = Note Ninja
 
 // The application's bundle identifier
 PRODUCT_BUNDLE_IDENTIFIER = com.example.codelab
