+++ web/index.html
@@ -29,7 +29,7 @@
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>codelab</title>
+  <title>Note Ninja</title>
   <link rel="manifest" href="manifest.json">
 
   <script>
